feat(keymap): update keymap to include HYPR key
The keymap configuration has been updated to uncomment the HYPR definition
and modify the default layer bindings to include the HYPR key. This change
allows for the use of the HYPR key in the default layer, providing additional
functionality to the user. The updated keymap includes changes to the sensor
bindings and other layers to accommodate the new HYPR key.